drm/virtio: add scaling support
The gem buffer created by i915 in guest vm may be
smaller than the resolution determined by the
mode setting operation. In that case, a new framebuffer
may be created by virtio gpu driver, and i915 rendering
must be utilized to do the composition. To save the
effort, pass the dst crtc info to the backend which will
scale the gem buffer to the crtc.

Tests done:
- Android Boot / reboot / recovery
- Android Boot using ACRN / fastboot
- Wifi / Wifi Hotspot
- BT / File Transfer
- Audio / Video Playback
- ADB
- Web Browsing
- Settings
- Type C check
- Factory reset

Tracked-On: OAM-130450
Signed-off-by: Hang Liu<hang1.liu@intel.com>
Signed-off-by: Zhao, Shirley <shirley.zhao@intel.com>